Combine the two Action type files
